tests/wpt/tests/css/css-values/urls/cross-origin/url-font-cross-origin-use-credentials.sub.html

<!DOCTYPE html>
<title>Request URL Modifiers: cross-origin(use-credentials) for @font-face</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#typedef-request-url-modifier-cross-origin-modifier">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@font-face {
  font-family: "TestFont";
  src: url("http://{{hosts[][]}}:{{ports[http][1]}}/fonts/Ahem.ttf?pipe=header(Access-Control-Allow-Origin,http://{{hosts[][]}}:{{ports[http][0]}})|header(Access-Control-Allow-Credentials,true)" cross-origin(use-credentials)) format("truetype");
}
.test { font-family: "TestFont" }
</style>
<div class="test">X</div>
<script>
promise_test(async t => {
  await document.fonts.ready;
  assert_true(document.fonts.check("16px TestFont"),
              "Cross-origin font with cross-origin(use-credentials) and ACAO+ACAC should load");
}, "cross-origin(use-credentials) allows cross-origin font loading with ACAO+ACAC");
</script>
